Part 4d final

+++ part4/4c/back/controllers/posts.js
@@ -36,9 +36,7 @@ const getTokenFrom = request => {
 
 postsRouter.post("/", async (request, response, next) => {
   const body = request.body
-  /* const token = tokenExtractor(request) */
   const token = getTokenFrom(request)
-  console.log(jwt.verify(token, process.env.SECRET))
   try {
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
@@ -67,9 +65,20 @@ postsRouter.post("/", async (request, response, next) => {
 })
 
 postsRouter.delete("/:id", async (request, response, next) => {
+  const token = getTokenFrom(request)
   try {
-    await Post.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" })
+    }
+    const user = await User.findById(decodedToken.id)
+    const post = await Post.findById(request.params.id)
+    if ( post.user.toString() === user._id.toString() ) {
+      await Post.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(403).end()
+    }
   } catch (exception) {
     next(exception)
   }
